GUI: Fix GCC Signed vs. Unsigned Comparison Warning

+++ gui/saveload-dialog.cpp
@@ -815,7 +815,7 @@ const Common::U32String &SaveLoadChooserGrid::getResultString() const {
 
 void SaveLoadChooserGrid::handleCommand(CommandSender *sender, uint32 cmd, uint32 data) {
 	const int slot = cmd + _curPage * _entriesPerPage - 1;
-	if (cmd <= _entriesPerPage && slot < _saveList.size()) {
+	if (cmd <= _entriesPerPage && slot < (int)_saveList.size()) {
 		activate(slot, Common::U32String());
 	}
